Old_ReAct_Agent/pyproject.toml
[project]
name = "react-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "black>=25.11.0",
    "isort>=7.0.0",
    "langchain>=1.1.2",
    "langchain-classic>=1.0.0",
    "langchain-openai>=1.1.0",
    "python-dotenv>=1.2.1",
]
